backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
google-cloud-firestore==2.11.0
google-auth==2.22.0
python-jose==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin==6.2.0
langchain
langchain-openai
langchain-core
pydantic==1.10.8
openai>=1.10.0,<2.0.0
python-dateutil==2.8.2
